backend/requirements.txt
django>=5.0,<5.3
djangorestframework>=3.15,<3.16
